@@ -1,6 +1,6 @@
 {
   "name": "boron",
-  "version": "0.1.2",
+  "version": "0.2.3",
   "description": "A collection of dialog animations with React.js",
   "main": "Boron.js",
   "author": "Yuanyan Cao",
@@ -10,12 +10,13 @@
     "url": "https://github.com/yuanyan/boron.git"
   },
   "dependencies": {
-    "react-kit": "^0.1.2"
+    "domkit": "^0.0.1"
   },
   "peerDependencies": {
-    "react": ">=0.12.0"
+    "react": ">=0.14.0"
   },
   "devDependencies": {
+    "react-dom": ">=0.14.0",
     "reactify": "^0.17.0",
     "browserify": "^6.3.3",
     "browserify-shim": "^3.8.0",
